Support routing through pages defined in sanity
